PSUH 12 stable

+++ src/vue/actualites/actualitesVue.php
@@ -2,11 +2,16 @@
     $title = 'Actualités';
     ob_start();
 ?>
-
-<div id='p-admin'>
-    <h2>Panneau administratif</h2>
-    <a href="index.php?action=formulaireActu">Nouvelle actualité</a>
-</div>
+    <h1 id='titrePage'>Actualités</h1>
+<?php 
+    if ($estAdmin === true) {
+?>
+        <div id='p-admin'>
+            <a href="index.php?action=formulaireActu"><img src="src/images/ajouter.svg" alt="Ajouter"><p>Nouvelle actualité</p></a>
+        </div>
+<?php 
+    }
+?>
 
 <div id='actualites'>
 <?php 
@@ -14,8 +19,10 @@
 ?>
     <div class="actu">
         <?php 
+        if ($estAdmin === true) {
             echo'<a id="editer" href="index.php?action=formulaireActu&id='. $actu->getId().'">Éditer <img src="src/images/editer.svg" alt="Éditer"></a>';
             echo'<a id="supprimer" href="index.php?action=supprimerActu&id='. $actu->getId() .'">Supprimer <img src="src/images/supprimer.svg" alt="Supprimer"></a>';
+        }
 
             echo"<div class='entete'>";
                 echo'<h1>'. $actu->getTitre() .'</h1>';
